Moved c&p functionality one level up to TableView, added clipboard icon. Keeps copy origin on table switch if origin is not a link cell. Displays clipboard icon when origin cell is set.
