+++ src/main/scss/header.scss
@@ -145,4 +145,14 @@ header {
       border-color: $color-primary-lighter;
     }
   }
+
+  #clipboard-icon {
+    margin-top: 10px;
+    display: inline-block;
+    float: right;
+    padding: 10px;
+    padding-right: 0;
+
+    pointer-events: none;
+  }
 }
